telegram_bot/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
aio-pika = "^9.4.1"
dependency-injector = "^4.41"
asyncpg = "^0.29"
sqlalchemy = "^2.0.28"
python-telegram-bot = {version="^21.3", extras=['job-queue', 'http2']}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'